Fixes based on code review. Move serialization logic back to Entry
