+++ src/main/java/com/configcat/ConfigFetcher.java
@@ -114,8 +114,8 @@ class ConfigFetcher implements Closeable {
             public void onResponse(@NotNull Call call, @NotNull Response response) {
                 try (ResponseBody body = response.body()) {
                     String fetchTime = response.headers().get("date");
-                    if (fetchTime == null || fetchTime.isEmpty() || CacheUtils.DateTimeUtils.isValidDate(fetchTime)) {
-                        fetchTime = CacheUtils.DateTimeUtils.format(System.currentTimeMillis());
+                    if (fetchTime == null || fetchTime.isEmpty() || DateTimeUtils.isValidDate(fetchTime)) {
+                        fetchTime = DateTimeUtils.format(System.currentTimeMillis());
                     }
                     if (response.isSuccessful() && body != null) {
                         String content = body.string();
