+++ src/test/java/com/configcat/EntrySerializationTest.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-public class CacheUtilTests {
+public class EntrySerializationTest {
 
     private static final String TEST_JSON = "{ f: { fakeKey: { v: %s, t: %s, p: [] ,r: [] } } }";
 
@@ -14,10 +14,10 @@ public class CacheUtilTests {
     public void serialize() {
         String json = String.format(TEST_JSON, "test", "1");
         Config config = Utils.gson.fromJson(json, Config.class);
-        String fetchTimeRaw = CacheUtils.DateTimeUtils.format(System.currentTimeMillis());
+        String fetchTimeRaw = DateTimeUtils.format(System.currentTimeMillis());
         Entry entry = new Entry(config, "fakeTag", json, fetchTimeRaw);
 
-        String serializedString = CacheUtils.serialize(entry);
+        String serializedString = entry.serialize();
 
         assertEquals(String.format(SERIALIZED_DATA, fetchTimeRaw, "fakeTag", json), serializedString);
     }
@@ -26,9 +26,9 @@ public class CacheUtilTests {
     public void deserialize() throws Exception {
         String json = String.format(TEST_JSON, "test", "1");
         long currentTimeMillis = System.currentTimeMillis();
-        String fetchTimeRaw = CacheUtils.DateTimeUtils.format(currentTimeMillis);
+        String fetchTimeRaw = DateTimeUtils.format(currentTimeMillis);
 
-        Entry entry = CacheUtils.deserialize(String.format(SERIALIZED_DATA, fetchTimeRaw, "fakeTag", json));
+        Entry entry = Entry.fromString(String.format(SERIALIZED_DATA, fetchTimeRaw, "fakeTag", json));
 
         assertNotNull(entry);
         assertEquals(entry.getFetchTimeRaw(), fetchTimeRaw);
@@ -40,24 +40,24 @@ public class CacheUtilTests {
 
     @Test
     public void deserializeMissingValue() throws Exception {
-        Entry deserializeNull = CacheUtils.deserialize(null);
+        Entry deserializeNull = Entry.fromString(null);
         assertTrue(deserializeNull.isEmpty());
-        Entry deserializeEmpty = CacheUtils.deserialize("");
+        Entry deserializeEmpty = Entry.fromString("");
         assertTrue(deserializeEmpty.isEmpty());
     }
 
     @Test
     public void deserializeWrongFormat() throws Exception {
-        Exception assertThrows = assertThrows(Exception.class, () -> CacheUtils.deserialize("value with no new line"));
+        Exception assertThrows = assertThrows(Exception.class, () -> Entry.fromString("value with no new line"));
         assertEquals("Number of values is fewer than expected.", assertThrows.getMessage());
 
-        assertThrows = assertThrows(Exception.class, () -> CacheUtils.deserialize("value with one \n new line"));
+        assertThrows = assertThrows(Exception.class, () -> Entry.fromString("value with one \n new line"));
         assertEquals("Number of values is fewer than expected.", assertThrows.getMessage());
     }
 
     @Test
     public void deserializeInvalidDate() throws Exception {
-        Exception assertThrows = assertThrows(Exception.class, () -> CacheUtils.deserialize(String.format(SERIALIZED_DATA, "invalid", "fakeTag", "json")));
+        Exception assertThrows = assertThrows(Exception.class, () -> Entry.fromString(String.format(SERIALIZED_DATA, "invalid", "fakeTag", "json")));
 
         assertEquals("Invalid fetch time: invalid", assertThrows.getMessage());
     }
@@ -65,8 +65,8 @@ public class CacheUtilTests {
     @Test
     public void deserializeInvalidETag() {
         long currentTimeMillis = System.currentTimeMillis();
-        String fetchTimeRaw = CacheUtils.DateTimeUtils.format(currentTimeMillis);
-        Exception assertThrows = assertThrows(Exception.class, () -> CacheUtils.deserialize(String.format(SERIALIZED_DATA, fetchTimeRaw, "", "json")));
+        String fetchTimeRaw = DateTimeUtils.format(currentTimeMillis);
+        Exception assertThrows = assertThrows(Exception.class, () -> Entry.fromString(String.format(SERIALIZED_DATA, fetchTimeRaw, "", "json")));
 
         assertEquals("Empty eTag value.", assertThrows.getMessage());
     }
@@ -74,12 +74,12 @@ public class CacheUtilTests {
     @Test
     public void deserializeInvalidJson() {
         long currentTimeMillis = System.currentTimeMillis();
-        String fetchTimeRaw = CacheUtils.DateTimeUtils.format(currentTimeMillis);
-        Exception assertThrows = assertThrows(Exception.class, () -> CacheUtils.deserialize(String.format(SERIALIZED_DATA, fetchTimeRaw, "fakeTag", "")));
+        String fetchTimeRaw = DateTimeUtils.format(currentTimeMillis);
+        Exception assertThrows = assertThrows(Exception.class, () -> Entry.fromString(String.format(SERIALIZED_DATA, fetchTimeRaw, "fakeTag", "")));
 
         assertEquals("Empty config jsom value.", assertThrows.getMessage());
 
-        assertThrows = assertThrows(Exception.class, () -> CacheUtils.deserialize(String.format(SERIALIZED_DATA, fetchTimeRaw, "fakeTag", "wrongjson")));
+        assertThrows = assertThrows(Exception.class, () -> Entry.fromString(String.format(SERIALIZED_DATA, fetchTimeRaw, "fakeTag", "wrongjson")));
 
         assertEquals("Invalid config JSON content: wrongjson", assertThrows.getMessage());
